+++ khd/kernels/add/add_tensor/cuda_implementation/__init__.py
@@ -1,1 +1,32 @@
-from .ops import add_tensor_cuda
+import torch
+
+from .ops import _add_tensor_forward_cuda, _add_tensor_forward_cuda_compilable, add_tensor_cuda
+
+
+class _AddTensor_CUDA(torch.autograd.Function):
+    @staticmethod
+    def forward(ctx, x: torch.Tensor, y: torch.Tensor) -> torch.Tensor:
+        if torch.compiler.is_compiling():
+            output = _add_tensor_forward_cuda_compilable(x, y)
+        else:
+            output = _add_tensor_forward_cuda(x, y)
+
+        return output
+
+    @staticmethod
+    def backward(ctx, output_grad: torch.Tensor) -> tuple[torch.Tensor | None]:
+        return output_grad, output_grad, None
+
+
+def add_tensor_cuda(x: torch.Tensor, y: torch.Tensor) -> torch.Tensor:
+    """tensor addition
+
+    Args:
+        x (torch.Tensor): input tensor
+        y (torch.Tensor): input tensor
+
+    Returns:
+        torch.Tensor: output tensor
+    """
+
+    return _AddTensor_CUDA.apply(x, y)
