+++ khd/kernels/add/add_tensor/cuda_implementation/ops.py
@@ -9,15 +9,27 @@ from ....utils import torch_custom_op
 _KERNEL_NAME = "add_tensor_forward_cuda"
 
 
-@AutoTune(configs=get_vectorized_autotune_configs(), trigger_keys=["x", "y", "dtype"])
-def _add_tensor_forward_cuda_autotuned(
-    x: torch.Tensor, y: torch.Tensor, dtype: torch.dtype, vectorized_loop_size: int, BLOCK_SIZE: int
+@AutoTune(configs=get_vectorized_autotune_configs(torch.float32), trigger_keys=["x", "y"])
+def _add_tensor_forward_cuda_autotuned_fp32(
+    x: torch.Tensor, y: torch.Tensor, vectorized_loop_size: int, BLOCK_SIZE: int
+) -> torch.Tensor:
+    return KernelRegistry.get_kernel(_KERNEL_NAME)(x, y, vectorized_loop_size, BLOCK_SIZE)
+
+
+@AutoTune(configs=get_vectorized_autotune_configs(torch.float16), trigger_keys=["x", "y"])
+def _add_tensor_forward_cuda_autotuned_fp16(
+    x: torch.Tensor, y: torch.Tensor, vectorized_loop_size: int, BLOCK_SIZE: int
 ) -> torch.Tensor:
     return KernelRegistry.get_kernel(_KERNEL_NAME)(x, y, vectorized_loop_size, BLOCK_SIZE)
 
 
 def _add_tensor_forward_cuda(x: torch.Tensor, y: torch.Tensor) -> torch.Tensor:
-    return _add_tensor_forward_cuda_autotuned(x, y, x.dtype)
+    if x.dtype in [torch.float16, torch.bfloat16]:
+        output = _add_tensor_forward_cuda_autotuned_fp16(x, y)
+    elif x.dtype == torch.float32:
+        output = _add_tensor_forward_cuda_autotuned_fp32(x, y)
+
+    return output
 
 
 @torch_custom_op(f"{LIBRARY_NAME}::{_KERNEL_NAME}", mutates_args={})
@@ -28,32 +40,3 @@ def _add_tensor_forward_cuda_compilable(x: torch.Tensor, y: torch.Tensor) -> tor
 @_add_tensor_forward_cuda_compilable.register_fake
 def _(x: torch.Tensor, y: torch.Tensor, vectorized_loop_size: int, BLOCK_SIZE: int) -> torch.Tensor:
     return torch.empty_like(x)
-
-
-class _AddTensor_CUDA(torch.autograd.Function):
-    @staticmethod
-    def forward(ctx, x: torch.Tensor, y: torch.Tensor) -> torch.Tensor:
-        if torch.compiler.is_compiling():
-            output = _add_tensor_forward_cuda_compilable(x, y)
-        else:
-            output = _add_tensor_forward_cuda(x, y)
-
-        return output
-
-    @staticmethod
-    def backward(ctx, output_grad: torch.Tensor) -> tuple[torch.Tensor | None]:
-        return output_grad, output_grad, None
-
-
-def add_tensor_cuda(x: torch.Tensor, y: torch.Tensor) -> torch.Tensor:
-    """tensor addition
-
-    Args:
-        x (torch.Tensor): input tensor
-        y (torch.Tensor): input tensor
-
-    Returns:
-        torch.Tensor: output tensor
-    """
-
-    return _AddTensor_CUDA.apply(x, y)
